Remove the event listener that selects newly created graphical model elements
Whenever a graphical model element is created (e.g., by parsing a textual expression), it is automatically selected by the diagram. This has the consequence that duplicate action event handlers get binded, i.e., the ones for the graphical model element and the ones for the embedded textual editor. To work around this limitation, the event listener that selects newly created graphical model elements is removed before parsing the textual expression and then added again when parsing has finished.